+++ api/@types/base-entity.d.ts
