+++ api/@types/promotions/entities.d.ts
