+++ api/@types/promotions/index.d.ts
