+++ api/constants/genders.ts
@@ -1,0 +1,1 @@
+export const USER_GENDER = ['OTHER', 'MALE', 'FEMALE'] as const;
