--- api/index.d.ts
@@ -1,12 +1,0 @@
-export * from './files';
-export * from './logs';
-export * from './permissions';
-export * from './promotions';
-export * from './roles';
-export * from './subscriptions';
-export * from './users';
-export * from './request';
-export * from './response';
-export * from './base-entity';
-
-export * from './i18n';
